test/chaos-pipeline/apps/recall/requirements.txt
torch==1.8.1+cpu
weaviate-client==2.4.0
nltk==3.5
